backend/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<3.16
mysqlclient>=2.2,<2.3
pytest>=8.3,<8.4
pytest-django>=4.9,<4.10
